+++ src/metadata.rs
@@ -4,7 +4,7 @@ use nom::{
     branch::alt,
     bytes::complete::take_while,
     character::complete::{char, satisfy, space1},
-    combinator::{iterator, map, recognize},
+    combinator::{iterator, recognize},
     sequence::preceded,
     Parser,
 };
@@ -23,27 +23,27 @@ use crate::{amount, empty_line, end_of_line, string, Currency, Decimal, IResult,
 /// "#;
 /// let beancount = beancount_parser_2::parse::<f64>(input).unwrap();
 /// let directive_metadata = &beancount.directives[0].metadata;
-/// assert_eq!(directive_metadata.get("title"), Some(&MetadataValue::String("Swiss Franc")));
+/// assert_eq!(directive_metadata.get("title"), Some(&MetadataValue::String("Swiss Franc".into())));
 /// ```
 #[derive(Debug, Clone, PartialEq)]
 #[non_exhaustive]
-pub enum Value<'a, D> {
+pub enum Value<D> {
     /// String value
-    String(&'a str),
+    String(String),
     /// A number or number expression
     Number(D),
     /// A [`Currency`]
     Currency(Currency),
 }
 
-pub(crate) fn parse<D: Decimal>(input: Span<'_>) -> IResult<'_, HashMap<&str, Value<'_, D>>> {
+pub(crate) fn parse<D: Decimal>(input: Span<'_>) -> IResult<'_, HashMap<&str, Value<D>>> {
     let mut iter = iterator(input, alt((entry.map(Some), empty_line.map(|_| None))));
     let map: HashMap<_, _> = iter.flatten().collect();
     let (input, _) = iter.finish()?;
     Ok((input, map))
 }
 
-fn entry<D: Decimal>(input: Span<'_>) -> IResult<'_, (&str, Value<'_, D>)> {
+fn entry<D: Decimal>(input: Span<'_>) -> IResult<'_, (&str, Value<D>)> {
     let (input, _) = space1(input)?;
     let (input, key) = recognize(preceded(
         satisfy(char::is_lowercase),
@@ -52,9 +52,9 @@ fn entry<D: Decimal>(input: Span<'_>) -> IResult<'_, (&str, Value<'_, D>)> {
     let (input, _) = char(':')(input)?;
     let (input, _) = space1(input)?;
     let (input, value) = alt((
-        map(string, Value::String),
-        map(amount::expression, Value::Number),
-        map(amount::currency, Value::Currency),
+        string.map(ToOwned::to_owned).map(Value::String),
+        amount::expression.map(Value::Number),
+        amount::currency.map(Value::Currency),
     ))(input)?;
     let (input, _) = end_of_line(input)?;
     Ok((input, (*key.fragment(), value)))